Move abstract-storage to storage
