+++ turo/lib/storage/abstract-storage.js
