+++ turo/lib/storage/app-bundle-storage.js
@@ -1,5 +1,5 @@
 import * as files from 'turo-documents';
-import { CompositeStorage, BundleDocumentLoader } from '../abstract-storage';
+import { CompositeStorage, BundleDocumentLoader } from '../storage/abstract-storage';
 
 export const bundleLoader = new BundleDocumentLoader(files);
 export const storage = new CompositeStorage([ bundleLoader ]);
