+++ turo/test/turo-fixture-storage.js
@@ -1,5 +1,5 @@
 import * as builtIn from 'turo-documents';
-import { CompositeStorage, BundleDocumentLoader } from '../lib/abstract-storage';
+import { CompositeStorage, BundleDocumentLoader } from '../lib/storage/abstract-storage';
 
 import importing from './fixtures/importing.turo';
 import fromDisk from './fixtures/loaded-from-disk.turo';
